p07_quantization/requirements.txt
torch>=2.1.0
transformers>=4.45.0
datasets>=2.20.0
accelerate>=0.34.0
huggingface_hub>=0.25.0
safetensors>=0.4.0
numpy>=1.24.0
tqdm>=4.65.0
auto-gptq>=0.7.0
autoawq>=0.2.0
bitsandbytes>=0.43.0
sentencepiece>=0.1.99
